Hacky solution to set the ambari-server user
It should not be necessary to modify the `ambari-server.properties` file
manually like this. The `ambari-server setup` already makes the proper
modifications to the properties file, but somehow they don't make it
into the docker layer.

Removing `setup-server` is just because the script will not work if you
call it a second time (the interaction is different on second/subsequent
calls to `ambari-server setup` because the setup is already partially
done).